solutions/aisearch-mcp4aoai/requirements.txt
mcp[cli]
openai
ipykernel
httpx
python-dotenv
azure-core
azure-search-documents
starlette
uvicorn
semantic-kernel